minor updates
- improved some parts of the documentation
- removed unused `symbol_levels` variable
- EoT symbols are defined with int8_t instead of floats
- new #defines for convenience

+++ encode/symbols.c
@@ -13,4 +13,4 @@ const int8_t symbol_map[4]={+1, +3, -1, -3};
 const int8_t symbol_list[4]={-3, -1, +1, +3};
 
 //End of Transmission symbol pattern
-const float eot_symbols[8]={+3, +3, +3, +3, +3, +3, -3, +3};
+const int8_t eot_symbols[8]={+3, +3, +3, +3, +3, +3, -3, +3};
